Adds a readable Request to script
Scripts now uses a ProcessRequest method that receives a Request type with method and target information. More information will be added later.

Some dubious coding decision have been made and flagged as TODO. The project needs to grow a bit more so that these decisions can be made with a better effort.