Started adding Query Parameters to the Angel Script
I got a bit lost when registering the types, and it seems I've made a series of bad decisions. It's hard for a BindRequest to be able to return a BindQueryParameters without needing extra memory or using new allocations.

Also, QueryParameters.Get() method should be GetValue, as Get alone is nonsensical.